Split sources to platform and logic
