+++ crates/lib/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod ffi;
 pub mod metrics;
+pub(crate) mod platform;
 pub mod sources;
