+++ crates/lib/src/platform/io_service.rs
@@ -1,0 +1,80 @@
+use std::mem::MaybeUninit;
+
+use core_foundation::{
+  base::{CFAllocatorRef, kCFAllocatorDefault},
+  dictionary::{CFDictionaryRef, CFMutableDictionaryRef},
+};
+
+use super::WithError;
+
+#[link(name = "IOKit", kind = "framework")]
+unsafe extern "C" {
+  fn IOServiceMatching(name: *const i8) -> CFMutableDictionaryRef;
+  fn IOServiceGetMatchingServices(mainPort: u32, matching: CFDictionaryRef, existing: *mut u32) -> i32;
+  fn IOIteratorNext(iterator: u32) -> u32;
+  fn IORegistryEntryGetName(entry: u32, name: *mut i8) -> i32;
+  fn IORegistryEntryCreateCFProperties(
+    entry: u32,
+    properties: *mut CFMutableDictionaryRef,
+    allocator: CFAllocatorRef,
+    options: u32,
+  ) -> i32;
+  fn IOObjectRelease(obj: u32) -> u32;
+}
+
+pub fn cfio_get_props(entry: u32, name: String) -> WithError<CFDictionaryRef> {
+  unsafe {
+    let mut props: MaybeUninit<CFMutableDictionaryRef> = MaybeUninit::uninit();
+    if IORegistryEntryCreateCFProperties(entry, props.as_mut_ptr(), kCFAllocatorDefault, 0) != 0 {
+      return Err(format!("Failed to get properties for {}", name).into());
+    }
+
+    Ok(props.assume_init())
+  }
+}
+
+pub struct IOServiceIterator {
+  existing: u32,
+}
+
+impl IOServiceIterator {
+  pub fn new(service_name: &str) -> WithError<Self> {
+    let service_name = std::ffi::CString::new(service_name).unwrap();
+    let existing = unsafe {
+      let service = IOServiceMatching(service_name.as_ptr() as _);
+      let mut existing = 0;
+      if IOServiceGetMatchingServices(0, service, &mut existing) != 0 {
+        return Err(format!("{} not found", service_name.to_string_lossy()).into());
+      }
+      existing
+    };
+    Ok(Self { existing })
+  }
+}
+
+impl Drop for IOServiceIterator {
+  fn drop(&mut self) {
+    unsafe {
+      IOObjectRelease(self.existing);
+    }
+  }
+}
+
+impl Iterator for IOServiceIterator {
+  type Item = (u32, String);
+
+  fn next(&mut self) -> Option<Self::Item> {
+    let next = unsafe { IOIteratorNext(self.existing) };
+    if next == 0 {
+      return None;
+    }
+
+    let mut name = [0; 128];
+    if unsafe { IORegistryEntryGetName(next, name.as_mut_ptr()) } != 0 {
+      return None;
+    }
+
+    let name = unsafe { std::ffi::CStr::from_ptr(name.as_ptr()) };
+    Some((next, name.to_string_lossy().to_string()))
+  }
+}
